requirements.txt
aiohttp           
astroid         
async-timeout     
attrs             
chardet           
colorama          
discord.py        
idna              
isort             
lazy-object-proxy 
mccabe            
multidict       
pip               
pylint            
pyperclip         
pypresence        
setuptools        
six               
toml              
wrapt             
yarl              